src/inference/requirements.txt
fastapi==0.115.12
mlflow==2.22.1
mlflow_skinny==2.22.1
numpy
pandas
pydantic==2.11.5
python-dotenv==1.1.0
python_box==7.3.2
PyYAML==6.0.2
starlette
uvicorn==0.34.3
xgboost